pyproject.toml
[project]
name = "cs50-ci-action"
version = "0.1.0"
description = "Testing pipeline for CS50 problem sets"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=46.0.2",
    "dotenv>=0.9.9",
    "pyjwt>=2.10.1",
    "requests>=2.32.5",
]

[dependency-groups]
dev = ["pytest>=8.4.1", "pytest-cov>=6.2.1", "ruff>=0.12.10"]

[tool.ruff]
line-length = 100
target-version = "py312"
preview = true

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "C90",   # mccabe
    "I",     # isort
    "ASYNC", # flake8-async
    "B",     # flake8-bugbear
    "BLE",   # flake8-blind-except
    "COM",   # flake8-commas
    "C4",    # flake8-comprehensions
    "DTZ",   # flake8-datetimez
    "EM",    # flake8-errmsg
    "ICN",   # flake8-import-conventions
    "PYI",   # flake8-pyi
    "Q",     # flake8-quotes
    "SIM",   # flake8-simplify
    "TCH",   # flake8-typechecking
    "PTH",   # flake8-use-pathlib
    "UP",    # pyupgrade
    "RUF",   # ruff-specific
    "TRY",   # tryceratops
    "FURB",  # refurb
]
ignore = ["COM812"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = ["--cov-report=term", "--cov=."]
pythonpath = "src"
